# src/DocumentManagementSystem.Api/Controllers/DocumentsController.cs
using DocumentManagementSystem.Application.Dtos;
using DocumentManagementSystem.Application.Services;
using Microsoft.AspNetCore.Mvc;

namespace DocumentManagementSystem.Api.Controllers;

[ApiController]
[Route("documents")]
public sealed class DocumentsController(IDocumentService service) : ControllerBase
{
    [HttpPost]
    [Consumes("multipart/form-data")]
    [ProducesResponseType<DocumentResponse>(StatusCodes.Status201Created)]
    [ProducesResponseType(StatusCodes.Status400BadRequest)]
    public async Task<ActionResult<DocumentResponse>> Upload(
        [FromForm] UploadDocumentRequest request,
        CancellationToken cancellationToken)
    {
        var result = await service.UploadAsync(request, cancellationToken);
        return CreatedAtAction(nameof(GetById), new { id = result.Id }, result);
    }

    [HttpGet]
    public ActionResult<IReadOnlyList<DocumentResponse>> GetAll() => Ok(service.GetAll());

    [HttpGet("{id:guid}")]
    public ActionResult<DocumentResponse> GetById(Guid id)
    {
        var document = service.Get(id);
        return document is null ? NotFound() : Ok(document);
    }

    [HttpGet("{id:guid}/content")]
    [ProducesResponseType(StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status404NotFound)]
    public IActionResult Download(Guid id)
    {
        var document = service.GetDownload(id);
        return document is null
            ? NotFound()
            : File(document.Content, document.ContentType, document.FileName);
    }

    [HttpDelete("{id:guid}")]
    public IActionResult Delete(Guid id) => service.Delete(id) ? NoContent() : NotFound();
}